Moves Book to components
